Made sure to pass longs in to curl_easy_setopt where necessary in the libtest code.

+++ tests/libtest/lib540.c
@@ -30,9 +30,9 @@ static void init(CURLM *cm, const char* url, const char* userpwd,
   curl_easy_setopt(eh, CURLOPT_URL, url);
   curl_easy_setopt(eh, CURLOPT_PROXY, PROXY);
   curl_easy_setopt(eh, CURLOPT_PROXYUSERPWD, userpwd);
-  curl_easy_setopt(eh, CURLOPT_PROXYAUTH, CURLAUTH_ANY);
-  curl_easy_setopt(eh, CURLOPT_VERBOSE, 1);
-  curl_easy_setopt(eh, CURLOPT_HEADER, 1);
+  curl_easy_setopt(eh, CURLOPT_PROXYAUTH, (long)CURLAUTH_ANY);
+  curl_easy_setopt(eh, CURLOPT_VERBOSE, 1L);
+  curl_easy_setopt(eh, CURLOPT_HEADER, 1L);
   curl_easy_setopt(eh, CURLOPT_HTTPHEADER, headers); /* custom Host: */
 
   curl_multi_add_handle(cm, eh);
