+++ tests/libtest/lib552.c
@@ -158,17 +158,17 @@ int test(char *URL)
     curl_easy_setopt(curl, CURLOPT_DEBUGFUNCTION, my_trace);
     curl_easy_setopt(curl, CURLOPT_DEBUGDATA, &config);
     /* the DEBUGFUNCTION has no effect until we enable VERBOSE */
-    curl_easy_setopt(curl, CURLOPT_VERBOSE, 1);
+    curl_easy_setopt(curl, CURLOPT_VERBOSE, 1L);
 
     /* setup repeated data string */
     for (i=0; i < sizeof data; ++i)
         data[i] = fill[i % sizeof fill];
 
     /* Post */
-    curl_easy_setopt(curl, CURLOPT_POST, 1);
+    curl_easy_setopt(curl, CURLOPT_POST, 1L);
 
     /* Setup read callback */
-    curl_easy_setopt(curl, CURLOPT_POSTFIELDSIZE, sizeof data);
+    curl_easy_setopt(curl, CURLOPT_POSTFIELDSIZE, (long) sizeof data);
     curl_easy_setopt(curl, CURLOPT_READFUNCTION, read_callback);
 
     /* Write callback */
@@ -182,7 +182,7 @@ int test(char *URL)
     curl_easy_setopt(curl, CURLOPT_URL, URL);
 
     /* Accept any auth. But for this bug configure proxy with DIGEST, basic might work too, not NTLM */
-    curl_easy_setopt(curl, CURLOPT_PROXYAUTH, CURLAUTH_ANY);
+    curl_easy_setopt(curl, CURLOPT_PROXYAUTH, (long)CURLAUTH_ANY);
 
     res = curl_easy_perform(curl);
     fprintf(stderr, "curl_easy_perform = %d\n", (int)res);
